+++ platform/src/app/auth/login/page.tsx
@@ -6,14 +6,21 @@ import { Button, Checkbox, FormControl, FormControlLabel } from '@mui/material';
 import TextField from '@mui/material/TextField';
 import Link from 'next/link';
 import { useRouter } from 'next/navigation';
-import { useState } from 'react';
+import { useEffect, useState } from 'react';
 
 export default function Login() {
     const { apiPOST } = useApi()
+    const user = useUserStore((state) => state.user )
     const setUser = useUserStore((state) => state.setUser )
     const setToken = useCSRFStore((state) => state.setToken)
 
     const router = useRouter()
+
+    useEffect(() => {
+        if (user) {
+            router.push('/')
+        }
+    }, [user])
 
     const [email, setEmail] = useState<string>('')
     const [password, setPassword] = useState<string>('');
@@ -30,11 +37,14 @@ export default function Login() {
             password: password
         })
         .then((res) => {
-            setUser(res)
-            router.push('/');
+            setUser(res.user)
             if (res.csrfToken) {
                 setToken(res.csrfToken);
             }
+        })
+        
+        .finally(() => {
+            router.push('/');
         })
     }
     
@@ -46,18 +56,21 @@ export default function Login() {
                     <h1>Login</h1>
                     <p className='note'><b>NOTE:</b> Currently the platform is in development, so account login/creation is disabled. This page does nothing... yet.</p>
                     <div className="form-content">
-                        <p className="instruction">Dont have an account? <Link href="/platform/auth/create">Create one here</Link></p>
+                        <p className="instruction">Dont have an account? <Link href="/auth/create">Create one here</Link></p>
                         <TextField
                           id="email"
                           label="Email"
                           value={email}
                           onChange={(e) => setEmail(e.target.value)}
+                          fullWidth
                         />
                         <TextField
                           id="password"
                           label="Password"
+                          type='password'
                           value={password}
                           onChange={(e) => setPassword(e.target.value)}
+                          fullWidth
                         />
                         <FormControlLabel
                             control={<Checkbox onChange={(e) => setRemember(e.target.checked)} />}
@@ -69,10 +82,11 @@ export default function Login() {
                             className='submit'
                             variant='contained'
                             onClick={() => login()}
+                            size='large'
                         >
                             Login
                         </Button>
-                        <p className='instruction'><Link href="/platform/auth/forgot">Forgot your password?</Link></p>
+                        <p className='instruction'><Link href="/auth/forgot">Forgot your password?</Link></p>
                     </div>
                 </div>
                 <div className="logo">